Fix weatherDetailed to newApi

+++ app/src/main/java/com/example/features/weather/detailed/screen/AdditionalInfoDay.kt
@@ -10,10 +10,10 @@ import androidx.compose.ui.unit.dp
 
 @Composable
 fun AdditionalInfoDay(
-    uvIndex: Float,
+    uvIndex: Double,
     humidity: Int,
-    wind: Float,
-    pressure: Float
+    wind: Double,
+    pressure: Int
 ) {
     Row(
         modifier = Modifier
