+++ app/src/main/java/com/example/features/weather/model/WeatherEvent.kt
@@ -3,6 +3,6 @@ package com.example.features.weather.model
 sealed class WeatherEvent {
 
     data object ToBack : WeatherEvent()
-    data class ToWeatherDetailed(val weatherId: Int) : WeatherEvent()
+    data class ToWeatherDetailed(val weatherId: Long) : WeatherEvent()
     data object LoadData : WeatherEvent()
 }
